Refactoring scanner and parser
