+++ src/ast/mod.rs
@@ -1,0 +1,26 @@
+mod ast;
+mod parse;
+mod scan;
+mod tokens;
+
+pub use ast::*;
+pub use tokens::*;
+
+pub fn get_ast(code: String) -> ast::Program {
+    println!("\n/// Scanning ///\n");
+
+    let mut scanner = scan::Scanner::new(code);
+    let tokens = scanner.scan();
+
+    for token in tokens.iter() {
+        print!("{}", token);
+    }
+    println!("");
+
+    println!("\n/// Parsing ///\n");
+
+    let mut parser = parse::Parser::new(tokens);
+    let ast_program = parser.parse();
+    println!("{}", ast_program);
+    ast_program
+}
