+++ src/ast/scan.rs
@@ -1,71 +1,8 @@
+use super::tokens::*;
 use crate::error::{ErrorHandler, Location};
 use std::collections::HashMap;
-use std::fmt;
 
 const RADIX: u32 = 10;
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum TokenType {
-    // Single character
-    LeftPar,
-    RightPar,
-    LeftBrace,
-    RightBrace,
-    Comma,
-    Dot,
-    Minus,
-    Plus,
-    Slash,
-    Star,
-    Percent,
-    Bang,
-    Equal,
-    Greater,
-    Less,
-    And,
-    Or,
-
-    // Two characters
-    BangEqual,
-    EqualEqual,
-    GreaterEqual,
-    LessEqual,
-    AndAnd,
-    OrOr,
-
-    // Literals
-    Identifier(String),
-    NumberLit(u64),
-    BooleanLit(bool),
-    // StringLit(String),
-
-    // Keywords
-    Let,
-    Var,
-    Fun,
-    If,
-    Else,
-    While,
-    Return,
-    True,
-    False,
-    Export,
-
-    // Other
-    SemiColon,
-    EOF,
-}
-
-pub struct Token {
-    pub t: TokenType,
-    pub loc: Location,
-}
-
-impl fmt::Display for Token {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{:?} ", self.t)
-    }
-}
 
 pub struct Scanner {
     error_handler: ErrorHandler,
